Basic react-native-video setup
