+++ package.json
@@ -11,7 +11,6 @@
   },
   "dependencies": {
     "@react-native-community/async-storage": "^1.11.0",
-    "@react-native-community/audio-toolkit": "^2.0.3",
     "@react-native-community/masked-view": "^0.1.10",
     "@react-native-community/slider": "^3.0.3",
     "@react-native-firebase/analytics": "^7.6.7",
@@ -29,6 +28,7 @@
     "react-native-screens": "^2.9.0",
     "react-native-splash-screen": "^3.2.0",
     "react-native-vector-icons": "^7.0.0",
+    "react-native-video": "^5.1.0-alpha8",
     "react-redux": "^7.2.1",
     "redux": "^4.0.5",
     "redux-saga": "^1.1.3"
