+++ src/components/Player/Player.js
@@ -1,60 +1,43 @@
-import React from 'react';
-import { View, Text, StyleSheet, TouchableOpacity } from 'react-native';
-import Icon from 'react-native-vector-icons/Ionicons';
-import { useNavigation } from '@react-navigation/native';
+import React, { createContext, useState } from 'react';
+import Video from 'react-native-video';
 
-import { theme } from '../../utils/theme';
-import Progress from './Progress';
+const initialState = {
+  playing: false,
+  paused: true,
+  url: '',
+  title: '',
+  duration: 0,
+  time: 0,
+  progress: 0,
+  play: () => {},
+};
 
-function Player() {
-  const { navigate } = useNavigation();
+const PlayerContext = createContext(initialState);
+
+function Player({ children }) {
+  const [playing, setPlaying] = useState(initialState.playing);
+  const [paused, setPaused] = useState(initialState.paused);
+  const [url, setUrl] = useState(initialState.url);
+  const [title, setTitle] = useState(initialState.title);
+  const [duration, setDuration] = useState(initialState.duration);
+  const [time, setTime] = useState(initialState.time);
+  const [progress, setProgress] = useState(initialState.progress);
+
+  function play({ url: episodeUrl, title: episodeTitle }) {
+    setPlaying(true);
+    setUrl(episodeUrl);
+    setTitle(episodeTitle);
+    setPaused(false);
+  }
 
   return (
-    <>
-      <Progress progress={1} />
-      <TouchableOpacity
-        activeOpacity={0.9}
-        onPress={() => navigate('PlayerModal')}
-        disabled={false}
-      >
-        <View style={styles.wrapper}>
-          <View style={styles.half}>
-            <Text style={{ ...styles.text, color: theme.fg }}>{'Nie odtwarzane'}</Text>
-          </View>
-          <View style={styles.half}>
-            <TouchableOpacity style={styles.controlButton} disabled={false}>
-              <Icon name={'play'} color={theme.fg} size={16} />
-            </TouchableOpacity>
-          </View>
-        </View>
-      </TouchableOpacity>
-    </>
+    <PlayerContext value={{ playing, paused, url, title, duration, time, progress, play }}>
+      {playing && (
+        <Video audioOnly={true} playInBackground={true} source={{ uri: url }} paused={paused} />
+      )}
+      {children}
+    </PlayerContext>
   );
 }
-
-const styles = StyleSheet.create({
-  wrapper: {
-    backgroundColor: theme.bg.medium,
-    borderBottomColor: theme.bg.dark,
-    borderBottomWidth: StyleSheet.hairlineWidth,
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-  },
-  half: {
-    flexDirection: 'row',
-  },
-  text: {
-    paddingVertical: 14,
-    paddingHorizontal: 10,
-    fontSize: 14,
-  },
-  controlButton: {
-    alignItems: 'center',
-    height: '100%',
-    paddingVertical: 14,
-    paddingHorizontal: 10,
-  },
-});
 
 export default Player;
