+++ src/components/TabBar/TabBar.js
@@ -2,12 +2,12 @@ import React from 'react';
 import { View } from 'react-native';
 import { BottomTabBar } from '@react-navigation/bottom-tabs';
 
-import Player from '../Player/Player';
+import TabBarPlayer from '../TabBarPlayer/TabBarPlayer';
 
 function TabBar(props) {
   return (
     <View>
-      <Player />
+      <TabBarPlayer />
       <BottomTabBar {...props} />
     </View>
   );
