+++ src/components/TabBarPlayer/Progress.js
