+++ src/store/sagas/sagas.js
@@ -1,16 +1,5 @@
-import { takeEvery, put, all } from 'redux-saga/effects';
-import SplashScreen from 'react-native-splash-screen';
-import { episodes as episodesAction } from '../actions/episodes';
-
-function* getEpisodes() {
-  const { episodes } = yield fetch('https://require.podcast.gq/episodes.json').then(res =>
-    res.json(),
-  );
-
-  yield put(episodesAction(episodes));
-
-  SplashScreen.hide();
-}
+import { takeEvery, all } from 'redux-saga/effects';
+import { getEpisodes } from './episodes';
 
 export function* rootSaga(): any {
   yield all([takeEvery('GET_EPISODES', getEpisodes)]);
